refactor(global api): renaming axiosinstance to api-client
